refactor: Modify Diagnosis result and history detail logic
Modify applying data to screen logic of Diagnosis result and history detail page

Related to: #32
