+++ client/core/navigation/src/main/java/com/peekaboo/navigation/NavRoutes.kt
@@ -23,6 +23,7 @@ sealed class NavRoutes(val route: String) {
 
     data object DiagnosisHistoryGraph: NavRoutes("diagnosis_history_graph")
     data object DiagnosisHistoryScreen: NavRoutes("diagnosis_history")
+    data object DiagnosisHistoryDetailScreen: NavRoutes("diagnosis_history_detail")
 
     data object DiagnosisQuickGraph: NavRoutes("diagnosis_quick_graph")
     data object DiagnosisQuickScreen: NavRoutes("diagnosis_quick")
