+++ client/feature/diagnosishistory/src/main/java/com/peekaboo/diagnosishistory/detail/DiagnosisHistoryDetailPageState.kt
@@ -1,0 +1,16 @@
+package com.peekaboo.diagnosishistory.detail
+
+import com.peekaboo.domain.entity.response.diagnosis.DiagnosisHistoryDetailModel
+import com.peekaboo.ui.base.PageState
+
+data class DiagnosisHistoryDetailPageState (
+    val diagnosisId: Int = 0,
+    val selectedDisease: String = "",
+    val customDescription: String = "",
+    val diseaseTotal: List<DiagnosisHistoryDetailModel.DiseaseDetailItem> = emptyList(),
+    val firstDiseaseModel: DiagnosisHistoryDetailModel.DiseaseDetailItem = DiagnosisHistoryDetailModel.DiseaseDetailItem(),
+    val secondDiseaseModel: DiagnosisHistoryDetailModel.DiseaseDetailItem = DiagnosisHistoryDetailModel.DiseaseDetailItem(),
+    val thirdDiseaseModel: DiagnosisHistoryDetailModel.DiseaseDetailItem = DiagnosisHistoryDetailModel.DiseaseDetailItem(),
+//    val diagnosisModel: DiagnosisConstModel = DiagnosisConstModel(),
+    val isDataUpdateSuccess: Boolean = false
+): PageState
